refactor(apps): move jsonrpc lib to unified place
